requirements.txt
djangorestframework
Pillow
pyjwt
python-decouple
django-cors-headers
ipykernel